Added javadoc messages to all events. Changed CurrentPlayerChangedEvent to pass type Player around instead of int. Changed EventManager to abstract with all static methods.

+++ core/src/edu/fresnostate/turnbased/event/CreateUnitEvent.java
@@ -2,13 +2,38 @@ package edu.fresnostate.turnbased.event;
 
 import edu.fresnostate.turnbased.UnitType;
 
+
+/**
+ * This event is send as a request to create a unit for the active player at the
+ * specified coordinates.
+ * 
+ * @author Daniel
+ *
+ */
 public class CreateUnitEvent implements Event
 {
-	public final UnitType unitType;
-	public final float	x;
-	public final float	y;
+	/**
+	 * The type of unit to create.
+	 */
+	public final UnitType	unitType;
+	/**
+	 * The x position of the unit to create.
+	 */
+	public final int		x;
+	/**
+	 * The y position of the unit to create.
+	 */
+	public final int		y;
 
-	public CreateUnitEvent (UnitType unitType, float x, float y)
+	/**
+	 * @param unitType
+	 *            The type of unit to create.
+	 * @param x
+	 *            The x position of the unit to create.
+	 * @param y
+	 *            The y position of the unit to create.
+	 */
+	public CreateUnitEvent (UnitType unitType, int x, int y)
 	{
 		this.unitType = unitType;
 		this.x = x;
